Add DOOM to mob_skill enum, update comment, consolidate
