+++ scripts/enum/mob_skill.lua
@@ -400,6 +400,8 @@ xi.mobSkill =
 
     HOWL                          = 1062,
 
+    DOOM                          = 1067,
+
     FRYPAN_1                      = 1081,
     SMOKEBOMB_1                   = 1082,
